@@ -6,7 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint .",
     "test": "vitest",
     "test:ci": "vitest run --reporter=dot"
   },
@@ -25,7 +25,7 @@
     "framer-motion": "^11.18.2",
     "highlight.js": "^11.11.1",
     "jotai": "^2.8.3",
-    "next": "15.2.3",
+    "next": "^16.0.3",
     "next-themes": "^0.4.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -42,13 +42,14 @@
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.19",
+    "@types/json-schema": "^7.0.15",
     "@types/node": "^20.14.11",
     "@types/react": "^19.0.1",
     "@types/react-dom": "^19.0.2",
-    "node-mocks-http": "^1.12.3",
     "autoprefixer": "^10.4.20",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^15.2.3",
+    "eslint": "^9.0.0",
+    "eslint-config-next": "^16.0.3",
+    "node-mocks-http": "^1.12.3",
     "postcss": "^8.4.47",
     "tailwindcss": "^3.4.10",
     "typescript": "^5.6.2",
