@@ -1,9 +1,29 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from dotenv import load_dotenv
-from backend.api.v1 import ai, agents
+import os
+import logging
+from .api.v1 import ai, agents
+
+# Configure logging
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
 
 load_dotenv()
+
+# Startup Check for Critical Environment Variables
+REQUIRED_ENV_VARS = [
+    "INTERNAL_API_KEY",
+    # "REPLICATE_API_TOKEN", # Optional depending on usage
+    # "OPENAI_API_KEY",      # Optional depending on usage
+]
+
+missing_vars = [var for var in REQUIRED_ENV_VARS if not os.getenv(var)]
+if missing_vars:
+    logger.warning(f"⚠️  MISSING ENVIRONMENT VARIABLES: {', '.join(missing_vars)}")
+    logger.warning("Some features may not work correctly.")
+else:
+    logger.info("✅ All critical environment variables found.")
 
 app = FastAPI(title="Karate Backend")
 
@@ -12,7 +32,7 @@ app.add_middleware(
     allow_origins=["http://localhost:3000", "https://*"],
     allow_credentials=True,
     allow_methods=["GET", "POST", "OPTIONS"],
-    allow_headers=["Authorization", "Content-Type", "x-api-key"],
+    allow_headers=["Authorization", "Content-Type", "x-api-key", "x-user-id"],
 )
 
 app.include_router(ai.router, prefix="/api/v1/ai", tags=["ai"]) 
@@ -25,4 +45,3 @@ def health():
 if __name__ == "__main__":
     import uvicorn
     uvicorn.run("backend.main:app", host="0.0.0.0", port=8000, reload=True)
-
