@@ -1,6 +1,4 @@
-// Local API shim using Convex's runtime proxies.
-// This avoids cross-dir ESM parsing issues while remaining compatible with useQuery/useMutation.
-import { anyApi, componentsGeneric } from 'convex/server';
-export const api = anyApi;
-export const internal = anyApi;
-export const components = componentsGeneric();
+// Use the generated API from the root convex folder
+import { api } from "../../../convex/_generated/api";
+export { api };
+export * from "../../../convex/_generated/api";
